refactor(visualization-block): move png download algorithm to a separate file
