+++ apps/web/src/utils/file.ts
@@ -19,3 +19,20 @@ export function readFile(
     }
   })
 }
+
+export function downloadFile(url: string, name: string) {
+  const downloadLink = document.createElement('a')
+
+  // some browsers don't support the a[download] solution
+  if (typeof downloadLink.download === 'undefined') {
+    // 'window.location.href = downloadLink' no long works due to some Chrome policies update
+    return
+  }
+
+  downloadLink.download = name
+  downloadLink.href = url
+
+  document.body.appendChild(downloadLink)
+  downloadLink.click()
+  document.body.removeChild(downloadLink)
+}
